LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/hit-test/z-index-expected.txt


PASS elementFromPoint should hit the last element in dom order.
FAIL elementFromPoint should hit the highest z-index element. assert_equals: expected Element node <div id="a" style="z-index: 5;"></div> but got Element node <div id="b"></div>
